refactor: collapse snapshot barrel into snapshot/index.ts
